refactor: move App to component directory
