+++ src/components/App.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import styled from "styled-components";
 import LazyLoad from "react-lazyload";
-import Map from "./components/Map";
-import Landing from "./components/landing/Landing";
-import Introduction from "./components/InfoCards";
+import Map from "./Map";
+import Landing from "./landing/Landing";
+import Introduction from "./InfoCards";
 import "@esri/calcite-components/dist/calcite/calcite.css";
 
 const lazyHeight = "70vh";
