+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-import App from "./App";
+import App from "./components/App";
 import { defineCustomElements } from "@esri/calcite-components/dist/custom-elements";
 
 defineCustomElements(window);
